+++ lib/providers/providers.dart
@@ -18,3 +18,4 @@ export 'device_and_app_info/device_and_app_info_provider.dart';
 export 'events/events_provider.dart';
 export 'search/search_provider.dart';
 export 'shared_preference/shared_preference_provider.dart';
+export 'notification/notification_provider.dart';
