+++ lib/views/root/root_page_view.dart
@@ -1,18 +1,14 @@
-import 'dart:io';
-
 import 'package:Medito/constants/constants.dart';
 import 'package:Medito/models/models.dart';
 import 'package:Medito/providers/providers.dart';
-import 'package:Medito/routes/routes.dart';
+import 'package:Medito/providers/root/root_combine_provider.dart';
 import 'package:Medito/services/notifications/notifications_service.dart';
 import 'package:Medito/views/player/player_view.dart';
-import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:Medito/views/player/widgets/mini_player_widget.dart';
+import 'package:Medito/widgets/widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:just_audio/just_audio.dart';
-import 'widgets/widgets.dart';
-import 'views/player/widgets/mini_player_widget.dart';
 
 class RootPageView extends ConsumerStatefulWidget {
   final Widget firstChild;
@@ -26,46 +22,10 @@ class RootPageView extends ConsumerStatefulWidget {
 class _RootPageViewState extends ConsumerState<RootPageView> {
   @override
   void initState() {
-    var audioPlayerProvider = ref.read(audioPlayerNotifierProvider);
-    audioPlayerProvider.initAudioHandler();
-    ref.read(remoteStatsProvider);
-    ref.read(postLocalStatsProvider);
-    ref.read(deviceAppAndUserInfoProvider);
-    ref.read(pageviewNotifierProvider).addListenerToPage();
-    _saveFcmTokenEvent(ref);
-    ref
-        .read(playerProvider.notifier)
-        .getCurrentlyPlayingTrack(isPlayAudio: false);
-    _checkNotificationPermission();
+    ref.read(rootCombineProvider(context));
+    ref.read(checkNotificationPermissionProvider(context));
     checkInitialMessage(ref);
-    var streamEvent = audioPlayerProvider.trackAudioPlayer.playerStateStream
-        .map((event) => event.processingState)
-        .distinct();
-    streamEvent.forEach((element) {
-      if (element == ProcessingState.completed) {
-        _handleAudioCompletion(ref);
-        _handleUserNotSignedIn(ref);
-      }
-    });
     super.initState();
-  }
-
-  void _checkNotificationPermission() {
-    Future.delayed(Duration(seconds: 4), () {
-      checkNotificationPermission().then((value) {
-        var checkPermissionStatusInLocalStorage = ref
-            .read(sharedPreferencesProvider)
-            .getBool(SharedPreferenceConstants.notificationPermission);
-        if (Platform.isAndroid &&
-            checkPermissionStatusInLocalStorage == null &&
-            value == AuthorizationStatus.denied) {
-          context.push(RouteConstants.notificationPermissionPath);
-        } else if (Platform.isIOS &&
-            value == AuthorizationStatus.notDetermined) {
-          context.push(RouteConstants.notificationPermissionPath);
-        }
-      });
-    });
   }
 
   @override
@@ -78,11 +38,10 @@ class _RootPageViewState extends ConsumerState<RootPageView> {
       var nextId = next?.audio.first.files.first.id;
       if (next != null &&
           (prev?.id != next.id || (prev?.id == next.id && prevId != nextId))) {
-        _handleAudioStartedEvent(
-          ref,
-          next.id,
-          next.audio.first.files.first.id,
-        );
+        ref.read(playerProvider.notifier).handleAudioStartedEvent(
+              next.id,
+              next.audio.first.files.first.id,
+            );
       }
     });
 
@@ -171,79 +130,5 @@ class _RootPageViewState extends ConsumerState<RootPageView> {
     } else {
       return widget.firstChild;
     }
-  }
-
-  void _saveFcmTokenEvent(
-    WidgetRef ref,
-  ) async {
-    var token = await requestGenerateFirebaseToken();
-    var fcm = SaveFcmTokenModel(fcmToken: token ?? '');
-    var event = EventsModel(
-      name: EventTypes.saveFcmToken,
-      payload: fcm.toJson(),
-    );
-    ref.read(eventsProvider(event: event.toJson()));
-  }
-
-  void _handleAudioStartedEvent(
-    WidgetRef ref,
-    String trackId,
-    String audioFileId,
-  ) {
-    var audio = AudioStartedModel(audioFileId: audioFileId, trackId: trackId);
-    var event = EventsModel(
-      name: EventTypes.audioStarted,
-      payload: audio.toJson(),
-    );
-    ref.read(eventsProvider(event: event.toJson()));
-  }
-
-  void _handleAudioCompletion(
-    WidgetRef ref,
-  ) {
-    final audioProvider = ref.read(audioPlayerNotifierProvider);
-    var extras = ref.read(audioPlayerNotifierProvider).mediaItem.value?.extras;
-    if (extras != null) {
-      _handleAudioCompletionEvent(
-        ref,
-        extras['fileId'],
-        extras['trackId'],
-      );
-      audioProvider.seekValueFromSlider(0);
-      audioProvider.pause();
-      ref.invalidate(packProvider);
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        ref.read(audioPlayPauseStateProvider.notifier).state =
-            PLAY_PAUSE_AUDIO.PAUSE;
-      });
-    }
-  }
-
-  void _handleUserNotSignedIn(WidgetRef ref) {
-    var _user = ref.read(authProvider.notifier).userRes.body as UserTokenModel;
-    if (_user.email == null) {
-      var params = JoinRouteParamsModel(screen: Screen.track);
-      context.push(
-        RouteConstants.joinIntroPath,
-        extra: params,
-      );
-    }
-  }
-
-  void _handleAudioCompletionEvent(
-    WidgetRef ref,
-    String audioFileId,
-    String trackId,
-  ) {
-    var audio = AudioCompletedModel(
-      audioFileId: audioFileId,
-      trackId: trackId,
-      updateStats: true,
-    );
-    var event = EventsModel(
-      name: EventTypes.audioCompleted,
-      payload: audio.toJson(),
-    );
-    ref.read(eventsProvider(event: event.toJson()));
   }
 }
