+++ lib/widgets/medito_webview_widget.dart
@@ -3,7 +3,6 @@ import 'package:Medito/views/main/app_bar_widget.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:webview_flutter/webview_flutter.dart';
